JuicedGeneral R Package

+++ makepackagehere.R
@@ -1,27 +1,46 @@
 # Make a package automatically.
-# Joost Kuckartz, jkuckartz1984@hotmail.com
-# modified: January 2021
-# version: 0.3
+
+# author: Joost Kuckartz, jkuckartz1984@hotmail.com
+# modified: July 2023
+# version: 1.0
 
 #########
 # USAGE #
 #########
-# 1. Modify variables in this file
-# 2. Source this file
-# 3. Edit 
-
-rm(list=ls())                           #clear memory
+# Before using this file, it's recommended you update existing packages.
+#
+# Step 1. Place this file plus "makepackagevariables.R" in the folder containing
+# the R files to make a package out of it, or place it in the directory
+# containing the folder named "R".
+#
+# Sample 1:
+# C:/
+# C:/code/               <-- place in this folder
+# C:/code/R/
+# C:/code/R/script1.R
+# C:/code/R/script2.R
+# 
+# Sample 2:
+# C:/
+# C:/code/               <-- place in this folder
+# C:/code/script1.R
+# C:/code/script2.R
+#
+# Step 2: Source this file
 
 #############
 # VARIABLES #
 #############
 
+rm(list=ls())
+
 # To edit the package variables, edit the 'makepackagevariables.R' file.
 # Keeping one file for each package is the most useful.
+miktexinstalled = FALSE        #set to true if you have MiKTeX installed to make a pdf manual
 
-################
-# MAKE PACKAGE #
-################
+##########
+# SCRIPT #
+##########
 
 #' Load a library. If it is not yet installed, it will install it.
 #' @param x The library to load
@@ -58,16 +77,50 @@ rm(list = c("loadlibrary"))                                  #no need for the fu
 
 #prepare move files settings
 copydelete=TRUE
-thisfiledir=getSrcDirectory(function(x) {x})                 #only works if this file is run or sourced!
-setwd(thisfiledir)
+curdir=getSrcDirectory(function(x) {x})                      #only works if this file is run or sourced!
+setwd(curdir)
 if (basename(getwd())=="R") {                                #if this directory is "R"
   if (file.exists("../DESCRIPTION")) {                       #and there is a DESCRIPTION file
-    setwd("../..")                                           #we already have a package directory structure
+    setwd("..")                                              #we already have a package directory structure
 	  copydelete=FALSE
   }
 }
+curdir=getwd()
 
+#load package variables
 source("makepackagevariables.R")
+
+#move source files to right directory
+newdir=paste0(curdir,"/R")
+dir.create(newdir, showWarnings = FALSE)
+if (copydelete) {
+  rfiles=list.files(pattern=".R", include.dirs=FALSE)       #list R files
+  locs=is.na(str_locate(rfiles,"makepackagehere")[,1])
+  rfiles=rfiles[locs]                                       #avoid makepackagehere file
+  locs=is.na(str_locate(rfiles,"makepackagevariables")[,1])
+  rfiles=rfiles[locs]                                       #avoid makepackagevariables file
+  locs=is.na(str_locate(rfiles,"installthispackage")[,1])
+  rfiles=rfiles[locs]                                       #avoid installthispackage file
+  message(paste0("Moving ",length(rfiles)," source files..."))
+  file.rename(paste0(curdir,"/",rfiles),paste0(newdir,"/",rfiles))    #move files to "./R" directory
+}
+
+#clean up existing package files for recreation (if present)
+isnewpackage=TRUE
+chkfile=paste0(curdir,"/DESCRIPTION")
+if (file.exists(chkfile)) {
+  isnewpackage=FALSE
+  file.remove(chkfile)                                    #remove DESCRIPTION file to update it
+}
+chkfile=paste0(curdir,"/NAMESPACE")
+if (file.exists(chkfile)) {
+  isnewpackage=FALSE
+  file.remove(chkfile)                                    #remove NAMESPACE file to update it
+}
+chkfile=paste0(curdir,"/",packagename,".Rproj")
+if (file.exists(chkfile)) {
+  file.remove(chkfile)                                    #remove Rproj file
+}
 
 #prepare description information
 if (length(packageauthors)==1) {
@@ -121,66 +174,36 @@ packagedesc=list(                                           #create package desc
 )
 
 #prepare package directory and settings and create it
-packagedir=paste0("./",packagename)
-if (!dir.exists(packagedir)) {                              #if directory doesn't exist yet
+if (isnewpackage) {                                         #if this is a new package
   message(paste0("Creating package ",packagename,"..."))
-  devtools::create(packagename, fields = packagedesc)
 } else {
   message(paste0("Updating package ",packagename,"..."))
-  chkfile=paste0(packagedir,"/DESCRIPTION")
-  if (file.exists(chkfile)) {
-    file.remove(chkfile)                                    #remove DESCRIPTION file to update it
-  }
-  chkfile=paste0(packagedir,"/NAMESPACE")
-  if (file.exists(chkfile)) {
-    file.remove(chkfile)                                    #remove NAMESPACE file to update it
-  }
-  chkfile=paste0(packagedir,"/",packagename,".Rproj")
-  if (file.exists(chkfile)) {
-    file.remove(chkfile)                                    #remove Rproj file to update it
-  }
-  
-  #trick to avoid question for nesting
-  challenge_nested_project <- function(path, name) return()
-  
-  rlang::env_unlock(env = asNamespace('usethis'))
-  rlang::env_binding_unlock(env = asNamespace('usethis'))
-  assign('challenge_nested_project', challenge_nested_project, envir = asNamespace('usethis'))
-  rlang::env_binding_lock(env = asNamespace('usethis'))
-  rlang::env_lock(asNamespace('usethis'))
-  
-  #create package 
-  usethis::create_package(packagedir,packagedesc,open=FALSE)
-  
-  copydelete=FALSE
 }
 
-#move source files to right directory
-if (copydelete) {
-  rfiles=list.files(pattern=".R", include.dirs=FALSE)       #list R files
-  locs=is.na(str_locate(rfiles,"makepackagehere")[,1])
-  rfiles=rfiles[locs]                                       #avoid makepackagehere file
-  locs=is.na(str_locate(rfiles,"makepackagevariables")[,1])
-  rfiles=rfiles[locs]                                       #avoid makepackagevariables file
-  message(paste0("Moving ",length(rfiles)," source files..."))
-  curdir=getwd()
-  newdir=paste0(curdir,"/",packagename,"/R")
-  file.rename(paste0(curdir,"/",rfiles),paste0(newdir,"/",rfiles))    #move files to "./packagename/R" directory
-}
+#trick to avoid question for nesting
+challenge_nested_project <- function(path, name) return()
+rlang::env_unlock(env = asNamespace('usethis'))
+rlang::env_binding_unlock(env = asNamespace('usethis'))
+assign('challenge_nested_project', challenge_nested_project, envir = asNamespace('usethis'))
+rlang::env_binding_lock(env = asNamespace('usethis'))
+rlang::env_lock(asNamespace('usethis'))
 
+#create package 
+usethis::create_package(curdir,packagedesc,open=FALSE)
+
+#install and document package
 message(paste0("Installing package ",packagename,"..."))
-#create documentation and install package
-setwd(paste0("./",packagename))                             #set package base folder as working directory
-document()                                                  #create documentation for all R files in the folder
-setwd("..")                                                 #return to original working directory
-detach_package(packagename,TRUE)                            #ensure the package is not loaded just in case.
-#install(packagename, keep_source=TRUE, dependencies = TRUE) #install this package and its dependencies
-install.packages(packagename, repos = NULL, dependencies = TRUE, type = "source") #install this package and its dependencies
+document(".")                                                #create documentation for all R files in the folder
+detach_package(packagename,TRUE)                             #ensure the package is not loaded just in case.
+install.packages(".", repos = NULL, dependencies = TRUE, type = "source") #install this package and its dependencies
 
 #make pdf documentation file
-message(paste0("Creating documentation pdf for package ",packagename,"..."))
-if (file.exists(paste0(packagename,".pdf"))) {
-  file.remove(paste0(packagename,".pdf"))
+if (miktexinstalled) {
+  message(paste0("Creating documentation pdf for package ",packagename,"..."))
+  if (file.exists(paste0(packagename,".pdf"))) {
+    file.remove(paste0(packagename,".pdf"))
+  }
+  path <- find.package(packagename)
+  system(paste(shQuote(file.path(R.home("bin"), "R")), "CMD", "Rd2pdf", shQuote(path)))
 }
-path <- find.package(packagename)
-system(paste(shQuote(file.path(R.home("bin"), "R")), "CMD", "Rd2pdf", shQuote(path)))
+message("All done! Best to close and reopen the R application.")
